Add SQLite database

+++ app/src/main/java/com/annasizova/loftcoin/util/ImgUrlFormat.java
@@ -5,5 +5,5 @@ import androidx.annotation.NonNull;
 public interface ImgUrlFormat {
 
     @NonNull
-    String format(int id);
+    String format(long id);
 }
